requirements.txt
Django==4.2.14
djangorestframework==3.15.2
pymupdf==1.24.8
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.4.0
dj-database-url==2.2.0
